requirements.txt
Babel==2.14.0
GitPython==3.1.42
Jinja2==3.1.3
MarkupSafe==2.0.1
Pillow==9.0.1
PySocks==1.7.1
PyWavelets==1.4.1
PyYAML==6.0
Pygments==2.11.2
Send2Trash==1.8.2
anyio==4.3.0
appdirs==1.4.4
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
async-lru==2.0.4
attrs==23.2.0
autopep8==2.1.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==6.1.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.2
conda-build==3.21.7
conda-content-trust==0+unknown
conda-package-handling==1.7.3
conda==4.11.0
contourpy==1.1.1
cryptography==36.0.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
docker-pycreds==0.4.0
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastjsonschema==2.19.1
filelock==3.4.2
fonttools==4.50.0
fqdn==1.5.1
fsspec==2024.10.0
ftfy==6.2.3
gitdb==4.0.11
glob2==0.7
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.26.2
idna==3.3
imageio==2.34.0
importlib_metadata==7.0.2
importlib_resources==6.3.1
ipykernel==6.29.3
ipython==7.31.1
isoduration==20.11.0
jedi==0.18.1
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kiwisolver==1.4.5
lazy_loader==0.3
libarchive-c==2.9
matplotlib-inline==0.1.2
matplotlib==3.7.5
mistune==3.0.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.21.2
nystrom-attention==0.0.11
opencv-python-headless==4.9.0.80
opencv-python==4.10.0.84
openpyxl==3.1.2
openslide-python==1.3.1
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==21.2.4
pkginfo==1.8.2
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.20
protobuf==4.25.3
psutil==5.8.0
ptyprocess==0.7.0
pyOpenSSL==21.0.0
pycodestyle==2.11.1
pycosat==0.6.3
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-json-logger==2.0.7
pytz==2021.3
pyzmq==25.1.2
referencing==0.34.0
regex==2024.9.11
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
ruamel-yaml-conda==0.15.100
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==1.42.0
setproctitle==1.3.3
setuptools==58.0.4
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.3.1
spams==2.6.5.4
staintools==2.1.2
terminado==0.18.1
threadpoolctl==3.3.0
tifffile==2023.7.10
timm==1.0.11
tinycss2==1.2.1
tokenizers==0.20.1
tomli==2.0.1
torch==1.11.0
torchelastic==0.2.2
torchstain==1.3.0
torchtext==0.12.0
torchvision==0.12.0
tornado==6.4
tqdm==4.62.3
traitlets==5.14.2
transformers==4.46.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wandb==0.16.4
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.37.1
zipp==3.18.1